--- apps/warehouse/src/app/pages/product/bases/index.ts
@@ -1,1 +1,0 @@
-export * from './base-product.entity';
