--- apps/warehouse/src/app/pages/product/components/index.ts
@@ -1,2 +1,0 @@
-export * from './custom-visible/visible-product.component';
-export * from './product-dialog/product-dialog.component';
