--- apps/warehouse/src/app/pages/product/container/index.ts
@@ -1,1 +1,0 @@
-export * from './product/product.component';
