--- apps/warehouse/src/app/pages/product/container/product/product.component.html
@@ -1,133 +1,0 @@
-<div class="d-flex mt-3 ml-2">
-  <div class='d-flex flex-column ml-3 border p-2' style='height: 80vh; width:250px; overflow-y: scroll'>
-    <nz-collapse>
-      <nz-collapse-panel
-        nzHeader='Tìm kiếm'
-        nzActive='true'>
-        <div [formGroup]='formGroup'>
-          <input nz-input class='form-control' type='search' formControlName='search'>
-        </div>
-      </nz-collapse-panel>
-    </nz-collapse>
-    <div>
-      <minhdu-fontend-collapse-radios
-        [formGroup]='formGroup' [controlName]="'inventoryType'" header='Tổng kho'
-        [active]="true"
-        [radios]='inventoryTitle'>
-      </minhdu-fontend-collapse-radios>
-    </div>
-    <div>
-      <minhdu-fontend-collapse-radios
-        [formGroup]='formGroup' [controlName]="'warehouseType'" header='Loại kho'
-        [active]="true"
-        [radios]='(warehouse$|async)||[]'>
-      </minhdu-fontend-collapse-radios>
-    </div>
-  </div>
-  <div class="d-flex flex-column ml-2" style="width: 80vw">
-    <div class="ml-2 d-flex justify-content-between">
-      <div class="font-2xl font-weight-bold">Sản phẩm</div>
-      <div>
-        <div>
-          <button nz-button nzType="primary" (click)='import()'>Nhập</button>
-          <button nz-button nzType="primary" class="ml-2">Xuất</button>
-          <button class='ml-2' nz-button nz-popover nzPopoverTitle='Ghim'
-                  [(nzPopoverVisible)]='visible'
-                  nzPopoverPlacement='bottomRight'
-                  [nzPopoverContent]='pinned'
-                  nzPopoverTrigger='click'>Ghim
-          </button>
-        </div>
-      </div>
-    </div>
-    <nz-table #tableDashBoard class="mt-2"
-              nzSize='small'
-              [nzData]="(products$|async)|| []"
-              [formGroup]='formGroup'
-              (nzPageIndexChange)="onPagination($event)"
-              [nzLoading]="(loading$|async)"
-              [nzScroll]="{ x: '2000px',y: '60vh'}"
-              [nzPageSize]="pageSizeTable"
-              *ngIf="(ui$|async) as ui"
-    >
-      <thead>
-      <tr>
-        <th [nzWidth]="'60px'" *ngIf="ui.stt.visible" [nzLeft]='ui.stt?.pinned||false'>STT</th>
-        <th *ngIf="ui.code.visible" [nzLeft]='ui.code?.pinned||false'>Mã hàng</th>
-        <th *ngIf="ui.name.visible" [nzLeft]='ui.name?.pinned||false'>Tên hàng</th>
-        <th *ngIf="ui.warehouseType.visible" [nzLeft]='ui.warehouseType?.pinned||false'>Loại kho</th>
-        <th *ngIf="ui.price.visible" [nzLeft]='ui.price?.pinned||false'>Đơn giá</th>
-        <th *ngIf="ui.amount.visible" [nzLeft]='ui.amount?.pinned||false'>Số lượng</th>
-        <th *ngIf="ui.totalCash.visible" [nzLeft]='ui.totalCash?.pinned||false'>Tổng tiền</th>
-        <th *ngIf="ui.barcode.visible" [nzLeft]='ui.barcode?.pinned||false'>Mã vạch</th>
-        <th *ngIf="ui.provider.visible" [nzLeft]='ui.provider?.pinned||false'>Nhà cung cấp</th>
-        <th *ngIf="ui.exp.visible" [nzLeft]='ui.exp?.pinned||false'>Hạn sử dụng</th>
-        <th *ngIf="ui.discount.visible" [nzLeft]='ui.discount?.pinned||false'>Chiếc khấu</th>
-        <!--          <th>-->
-        <!--            <div class='border-top text-center'>Số chứng từ</div>-->
-        <!--            <input disabled>-->
-        <!--          </th>-->
-        <th *ngIf="ui.unit.visible" [nzLeft]='ui.unit?.pinned||false'>Đơn vị</th>
-        <th *ngIf="ui.createdAt.visible" [nzLeft]='ui.createdAt?.pinned||false'>Ngày nhập kho</th>
-      </tr>
-      </thead>
-      <tbody>
-      <ng-template ngFor let-product [ngForOf]="tableDashBoard.data" let-i='index'>
-        <tr class="pointer" (contextmenu)='child.onContextMenu($event,product)' (click)='onUpdate(product)'>
-          <td *ngIf="ui.stt.visible" [nzLeft]='ui.stt?.pinned||false'>
-            {{(i + 1) + ((tableDashBoard.nzPageIndex - 1) * 10)}}
-          </td>
-          <td *ngIf="ui.code.visible" [nzLeft]='ui.code?.pinned||false'>
-            {{product?.code}}
-          </td>
-          <td *ngIf="ui.name.visible" [nzLeft]='ui.name?.pinned||false'>
-            {{product?.name}}
-          </td>
-          <td *ngIf="ui.warehouseType.visible" [nzLeft]='ui.warehouseType?.pinned||false'>
-            {{product.warehouse?.name}}
-          </td>
-          <td *ngIf="ui.price.visible" [nzLeft]='ui.price?.pinned||false'>
-            {{product.price | number:'1.0'|notEmpty:'0'}} đ
-          </td>
-          <td *ngIf="ui.amount.visible" [nzLeft]='ui.amount?.pinned||false'>
-            {{product?.amount}}
-          </td>
-          <td *ngIf="ui.totalCash.visible" [nzLeft]='ui.totalCash?.pinned||false' style='color: #4dbd74'>
-            {{(product.price * product.amount) - ((product.price + product.amount) * (product?.discount || 0)) | number:'1.0'|notEmpty:'0'}}
-          </td>
-          <td *ngIf="ui.barcode.visible" [nzLeft]='ui.barcode?.pinned||false'>
-            {{product?.barcode}}
-          </td>
-          <td *ngIf="ui.provider.visible" [nzLeft]='ui.provider?.pinned||false'>
-            {{product?.provider?.name}}
-          </td>
-          <td *ngIf="ui.exp.visible" [nzLeft]='ui.exp?.pinned||false'>
-            {{(product?.exp | date: 'dd/MM/yyyy') || 'Chưa có thông tin'}}
-          </td>
-          <td *ngIf="ui.discount.visible" [nzLeft]='ui.discount?.pinned||false' style='color: #4dbd74'>
-            {{(product.discount || 0) * 100 + '%' }}
-          </td>
-          <!--          <th class='pointer'>-->
-          <!--            {{product?.billCode}}-->
-          <!--          </th>-->
-          <td *ngIf="ui.unit.visible" [nzLeft]='ui.unit?.pinned||false'>
-            {{product.unit | transform:medicineConstant}}
-          </td>
-          <td *ngIf="ui.stt.visible" [nzLeft]='ui.stt?.pinned||false'>
-            {{(product?.createdAt | date: 'dd/MM/yyyy') || 'Đang cập nhật'}}
-          </td>
-        </tr>
-      </ng-template>
-      </tbody>
-    </nz-table>
-
-  </div>
-</div>
-<ng-template #pinned>
-  <minhdu-fontend-pinned-product></minhdu-fontend-pinned-product>
-</ng-template>
-<app-mouse-right #child
-                 (onAdd)='import()'
-                 (onDel)='onDelete($event)'
-                 (onUpdate)='onUpdate($event)'></app-mouse-right>
-
