--- apps/warehouse/src/app/pages/product/dto/index.ts
@@ -1,3 +1,0 @@
-export * from './add-product.dto';
-export * from './search-product.dto';
-export * from './update-product.dto';
