--- apps/warehouse/src/app/pages/product/dto/search-product.dto.ts
@@ -1,6 +1,0 @@
-import { AddProductDto } from './add-product.dto';
-
-export interface SearchProductDto extends Partial<AddProductDto> {
-  readonly skip: number,
-  readonly take: number,
-};
