--- apps/warehouse/src/app/pages/product/dto/update-product.dto.ts
@@ -1,7 +1,0 @@
-import {BaseUpdateDto} from "@minhdu-fontend/base-dto";
-import {BaseProductEntity} from "../bases";
-export interface BaseUpdateProductDto extends BaseProductEntity{
-
-}
-
-export type UpdateProductDto = BaseUpdateDto<BaseUpdateProductDto>
