--- apps/warehouse/src/app/pages/product/entities/index.ts
@@ -1,2 +1,0 @@
-export * from './product.entity';
-export * from './product-visible.entity';
