--- apps/warehouse/src/app/pages/product/entities/product-visible.entity.ts
@@ -1,17 +1,0 @@
-import {VisibleEntity} from "@minhdu-fontend/data-models";
-
-export interface ProductVisibleEntity {
-  readonly stt: VisibleEntity;
-  readonly code: VisibleEntity;
-  readonly name: VisibleEntity;
-  readonly warehouseType: VisibleEntity;
-  readonly price: VisibleEntity;
-  readonly amount: VisibleEntity;
-  readonly totalCash: VisibleEntity;
-  readonly barcode: VisibleEntity;
-  readonly provider: VisibleEntity;
-  readonly discount: VisibleEntity;
-  readonly exp: VisibleEntity;
-  readonly unit: VisibleEntity;
-  readonly createdAt: VisibleEntity;
-}
