--- apps/warehouse/src/app/pages/product/entities/product.entity.ts
@@ -1,9 +1,0 @@
-import { Branch } from '@minhdu-fontend/data-models';
-import { WarehouseUnit } from '../../warehouse/enums';
-import { WarehouseEntity } from '../../warehouse/entities';
-import { ProviderEntity } from '../../provider/entities';
-import { BaseProductEntity } from '../bases';
-
-export interface Product extends BaseProductEntity {
-
-}
