--- apps/warehouse/src/app/pages/product/product-routing.module.ts
@@ -1,17 +1,0 @@
-import {NgModule} from '@angular/core';
-import {RouterModule, Routes} from '@angular/router';
-import {ProductComponent} from "./container/product/product.component";
-
-const routes: Routes = [
-  {
-    path: '',
-    component: ProductComponent,
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class ProductRoutingModule {
-}
