--- apps/warehouse/src/app/pages/product/services/index.ts
@@ -1,1 +1,0 @@
-export * from './product.service';
