+++ apps/warehouse/src/app/pages/stock/bases/index.ts
@@ -1,0 +1,1 @@
+export * from './base-stock.entity';
