+++ apps/warehouse/src/app/pages/stock/components/index.ts
@@ -1,0 +1,2 @@
+export * from './custom-visible/visible-product.component';
+export * from './stock-dialog/stock-dialog.component';
