+++ apps/warehouse/src/app/pages/stock/components/stock-dialog/stock-dialog.component.html
