+++ apps/warehouse/src/app/pages/stock/constants/index.ts
