+++ apps/warehouse/src/app/pages/stock/constants/inventory-title.constant.ts
