+++ apps/warehouse/src/app/pages/stock/container/index.ts
@@ -1,0 +1,1 @@
+export * from './stock/stock.component';
