+++ apps/warehouse/src/app/pages/stock/dto/add-stock.dto.ts
@@ -1,0 +1,8 @@
+import {BaseStockEntity} from "../bases";
+import {BaseAddDto} from "@minhdu-fontend/base-dto";
+
+interface BaseAddStockDto extends BaseStockEntity {
+
+}
+
+export type AddStockDto = BaseAddDto<BaseAddStockDto>
