+++ apps/warehouse/src/app/pages/stock/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './add-stock.dto';
+export * from './search-stock.dto';
+export * from './update-stock.dto';
