+++ apps/warehouse/src/app/pages/stock/dto/search-stock.dto.ts
@@ -1,0 +1,6 @@
+import { AddStockDto } from './add-stock.dto';
+
+export interface SearchStockDto extends Partial<AddStockDto> {
+  readonly skip: number,
+  readonly take: number,
+};
