+++ apps/warehouse/src/app/pages/stock/dto/update-stock.dto.ts
@@ -1,0 +1,8 @@
+import {BaseStockEntity} from "../bases";
+import {BaseUpdateDto} from "@minhdu-fontend/base-dto";
+
+interface BaseUpdateStockDto extends BaseStockEntity{
+
+}
+
+export type UpdateStockDto = BaseUpdateDto<BaseUpdateStockDto>
