+++ apps/warehouse/src/app/pages/stock/entities/index.ts
@@ -1,0 +1,2 @@
+export * from './stock.entity';
+export * from './stock-visible.entity';
