+++ apps/warehouse/src/app/pages/stock/entities/stock.entity.ts
@@ -1,0 +1,9 @@
+import { Branch } from '@minhdu-fontend/data-models';
+import { WarehouseUnit } from '../../warehouse/enums';
+import { WarehouseEntity } from '../../warehouse/entities';
+import { ProviderEntity } from '../../provider/entities';
+import { BaseStockEntity } from '../bases';
+
+export interface Product extends BaseStockEntity {
+
+}
