+++ apps/warehouse/src/app/pages/stock/services/index.ts
@@ -1,0 +1,1 @@
+export * from './stock.service';
