+++ apps/warehouse/src/app/pages/stock/stock-routing.module.ts
@@ -1,0 +1,17 @@
+import {NgModule} from '@angular/core';
+import {RouterModule, Routes} from '@angular/router';
+import {StockComponent} from "./container/stock/stock.component";
+
+const routes: Routes = [
+  {
+    path: '',
+    component: StockComponent,
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class StockRoutingModule {
+}
